requirements.txt
python_version >= '3.6'
ccxt
python-binance
asyncio
